Etape4

+++ Caddie.cpp
@@ -28,6 +28,7 @@ void handlerSIGALRM(int sig);
 
 int main(int argc,char* argv[])
 {
+  printf("------------Lancement du processus caddie---------------\n");
   // Masquage de SIGINT
   sigset_t mask;
   sigaddset(&mask,SIGINT);
@@ -44,27 +45,24 @@ int main(int argc,char* argv[])
     exit(1);
   }
 
-  // Connexion à la base de donnée
-  connexion = mysql_init(NULL);
-  if (mysql_real_connect(connexion,"localhost","Student","PassStudent1_","PourStudent",0,0,0) == NULL)
-  {
-    fprintf(stderr,"(CADDIE) Erreur de connexion à la base de données...\n");
-    exit(1);  
-  }
-  fprintf(stderr,"\n(CADDIE %d) apres connexion a la BD\n", getpid());
-
   MESSAGE m;
   MESSAGE reponse;
+  MESSAGE requete;
   
-  char requete[200];
   char newUser[20];
-  MYSQL_RES  *resultat;
-  MYSQL_ROW  Tuple;
 
-  fprintf(stderr,"\n(CADDIE %d) avant la recuperation du pipe\n", getpid());
   // Récupération descripteur écriture du pipe
+  //fdWpipe = atoi(argv[1]);
+  if (argc < 2) {
+    fprintf(stderr, "(CADDIE %d) Aucun argument fourni pour fdWpipe\n", getpid());
+    exit(1);
+  }
+
   fdWpipe = atoi(argv[1]);
-  fprintf(stderr,"\n(CADDIE %d) apres la recuperation du pipe\n", getpid());
+  if (fdWpipe <= 0) {
+      fprintf(stderr, "(CADDIE %d) Descripteur de pipe invalide : %s\n", getpid(), argv[1]);
+      exit(1);
+  }
 
   while(1)
   {
@@ -72,66 +70,67 @@ int main(int argc,char* argv[])
     if (msgrcv(idQ,&m,sizeof(MESSAGE)-sizeof(long),getpid(),0) == -1)
     {
       perror("(CADDIE) Erreur de msgrcv");
+      sleep(3);
       exit(1);
     }
 
     switch(m.requete)
     {
       case LOGIN :  {
+                      pidClient = m.expediteur;
                       fprintf(stderr,"(CADDIE %d) Requete LOGIN reçue de %d\n",getpid(),m.expediteur);
                       break;
                     }
       case LOGOUT :   
                     {
                       mysql_close(connexion);
-                      exit(0);
+                      close(fdWpipe);
                       fprintf(stderr,"(CADDIE %d) Requete LOGOUT reçue de %d\n",getpid(),m.expediteur);
+                      exit(0);
                       break;
                     }
       case CONSULT :  
                     {
-                      reponse.type = m.expediteur;
-                      reponse.expediteur = getpid();
-                      reponse.requete = CONSULT;
-
-                      // Construction de la requête SELECT
-                      sprintf(requete, "SELECT * FROM UNIX_FINAL WHERE id = %d;", m.data1);
-
-                      // Exécution de la requête
-                      if (mysql_query(connexion, requete)) {
-                          fprintf(stderr, "Erreur lors de la requête SELECT : %s\n", mysql_error(connexion));
-                          mysql_close(connexion);
-                          exit(1);
-                      }
-
-                      // Récupération des résultats
-                      resultat = mysql_store_result(connexion);
-                      if (resultat == NULL) {
-                          fprintf(stderr, "Erreur lors de la récupération des résultats : %s\n", mysql_error(connexion));
-                          mysql_close(connexion);
-                          exit(1);
-                      }
-
-                      // Affichage des résultats
-                      int num_champs = mysql_num_fields(resultat);
-
-                      if ((Tuple = mysql_fetch_row(resultat))) 
+                      requete.type = m.expediteur;
+                      requete.expediteur = getpid();
+                      requete.requete = CONSULT;
+                      requete.data1 = m.data1;
+                      if (write(fdWpipe, &requete, sizeof(MESSAGE)) != sizeof(MESSAGE)) 
                       {
-                        reponse.data1 = atoi(Tuple[0]); //id
-                        strncpy(reponse.data2, Tuple[1], sizeof(reponse.data2)); //nom de l'article
-                        reponse.data5 = atof(Tuple[2]); //prix unitaire
-                        strncpy(reponse.data3, Tuple[3], sizeof(reponse.data3)); //stock
-                        strncpy(reponse.data4, Tuple[4], sizeof(reponse.data2)); //chemin de l'image
+                        perror("Erreur de write(1) dans le caddie"); 
+                        sleep(3);
+                        exit(1); 
                       }
 
-                      if(msgsnd(idQ, &reponse, sizeof(MESSAGE) - sizeof(long), 0) == -1)
+                      if (msgrcv(idQ,&requete,sizeof(MESSAGE)-sizeof(long),getpid(),0) == -1)
                       {
-                        perror("Erreur de msgsnd CONSULT de caddie\n");
+                        perror("(CADDIE) Erreur de msgrcv dans le CONSULT du caddie\n");
+                        sleep(3);
                         exit(1);
                       }
-                      kill(m.expediteur, SIGUSR1);
-                      fprintf(stderr,"Mise a jour envoyé\n");
 
+                      if(requete.data1 != -1)
+                      {
+                        reponse.type = pidClient;
+                        reponse.expediteur = getpid();
+                        reponse.requete = CONSULT;
+
+                        reponse.data1 = requete.data1; //id
+                        strncpy(reponse.data2, requete.data2, sizeof(reponse.data2)); //nom de l'article
+                        strncpy(reponse.data3, requete.data3, sizeof(reponse.data3)); //stock
+                        strncpy(reponse.data4, requete.data4, sizeof(reponse.data2)); //chemin de l'image
+                        reponse.data5 = requete.data5; //prix unitaire
+
+                        if(msgsnd(idQ, &reponse, sizeof(MESSAGE) - sizeof(long), 0) == -1)
+                        {
+                          perror("Erreur de msgsnd CONSULT de caddie\n");
+                          sleep(3);
+                          exit(1);
+                        }
+                        kill(m.expediteur, SIGUSR1);
+                        fprintf(stderr,"Mise a jour des consultation envoyé par le Caddie\n");
+                      }
+                      
                       fprintf(stderr,"(CADDIE %d) Requete CONSULT reçue de %d\n",getpid(),m.expediteur);
                       break;
                     }
